Upgrade Z3 version
The last PR (https://github.com/diem/diem/pull/9807) had to be reverted because the newer version of Z3 (with glibc 2.31) is not compatible with the current CI test setup (with Debian 10).

This PR update `dev-setup.sh` to install the newest Z3 (4.8.13) built with `glibc-2.28` so that it's compatible to the current CI setup.

Todo:
- `NetworkIdentity::add_members_interal` is disabled for proving because
it fails with Z3 4.8.13

Closes: #9962