[utils] make normalized module serializable
This is useful for turning key parts of on-chain module bytecode into a structured format that can be used by clients. For example: a validator or indexing service might expose an API that returns a JSON representation of a normalized module that a client can use to discover public functions in that module.

Closes: #46