[evm] Generalizing placeholder definitions & memory guard
This generalizes the placeholder mechanism (`${MAX_U12}` et. al) to the whole code emission. That is, wherever an `emitln!` or an `emit!` is used, placeholders will be substituted. The definition of placeholders also becomes more declarative. Finally, placeholder syntax now requires braces (`${X}` instead of `$X`).

The first application is to emit the memoryguard instruction.

Closes: #49