[move-toy-yul] Native function generators.
This PR wires up a design to define generators for native functions (new module `native_functions.rs`). It implements one of those functions, namely `Evm::sign`, which is needed to create tests for resources.

This function is used in a first resource test which, however, does not yet pass in this PR, and is handled with in a followup.

Closes: #82