[move-to-yul] Setup experiments.
This PR sets up experiments for code generation and wires them to the test infra. This will be needed for the upcoming control flow graph optimizations, and down the road also for other things.

A new module `experiments.rs` is added which declares constants for experiments. An experiment can be activated on the command line with `--experiment=<name>`. In a baseline tests, one can add the comment
```
// experiment: <name>
```

... to the source. The test is then run both with and without the experiment, producing different baseline files.

A first experiment is used to supress generation of source info in baseline tests by default. This is useful because a lot of useless diffs are created if source positions only slightly shift. Source info will now only be generated during tests if `--experiment=capture-source-info` is set. Only one of our tests has this on, in order to ensure source info keeps being tested.

Closes: #88