[move-vm] Reduce VALUE_DEPTH_MAX to 128 + enable release build for CLI
PR #95 is experiencing stack overflows in some unrelated tests. After some analysis, it appears that is caused by some stress tests in `BCSTests` and `EventTests` which construct (via generics) types which have a nesting depth of 256 and are excercising the VALUE_DEPTH_MAX constant in the VM. In debug builds this constants value is simply to high, as it triggers the recursive computation of a TypeLayout.

This PR reduces the size of VALUE_DEPTH_MAX to 128 and adopts the existing tests accordingly. A max depth of 128 seems more than sufficient. (It means, in practice, that a user cannot nest more than 128 struct types.)

The PR also enables release builds of the Move CLI and therefore unit tests. This will allow in the future to analyze such failures, as it is well known that Rust in debug mode has huge stack consumption. However, it does not enable (yet) CI testing in release mode, because unfortunately, this would disable all kind of `debug_assert!` in our code. (We should eventually make all such assertions available also in release builds.)

Closes: #100