[move-prover] Fixed bugs in shift
Issue #147 reported a bug in the left shift operation where the prover
failed to report an obviously violated postcondition.  The report also
observed that the shift operators were only for a few shift amounts.
It is unfortunate that the unimplemented shift amounts did not result
in a false error report.  Upon inspection of the code, I also found a
"TODO" that shift left doesn't delete bits that are shifted left of
the end of the destination.  In the prover, this would result in cases
where the unbounded integer value would exceed the upper bound imposed
by the bit width of the destination.

The use of a recursive operator raises a risk of non-termination, but
it seems to work fine in test cases, including a variable shift
(somewhat to my surprise).

I made the following changes and improvements
1. In prelude.bpl, implemented a recursive function $pow
   to raise an integer value an integer exponent. It explicitly
   returns and undefined result when the exponent is negative,
   to eliminate an obvious possibility for non-termination.
2. Change $shl and $shr to call this function to raise 2 to
   the shift amount, the multiply or divide the shifted value by
   that amount.
3. Added sized shift-left instructions $ShlU8/64/128 which call
   $shl and then mod the result by the appropriate amount to delete
   bits that would be shifted beyond the limits of the destination.
   No such change was required for $Shr, so there is only one $Shr
   operator.
4. Modified bytecode_translator to generation $ShlU8/64/128 instead of
   $Shl.
5. Added function tests for shift operators in shift.move
6. Excluded CVC5 from the new shift test and from fixed_point_arithm.move
   because it goes into infinite loops on shifts (because it has trouble
   with recursive functions like $pow, I think).
7. Redid some exp files because of trivial changes in output.
8. In move-stdlib/Errors.move, there was use of << in a spec that
   did not account for bits lost off the left end. In my opinion, this
   is a (tiny) error in Errors::make, but I didn't know how to fix it,
   so I did the shift modulo 2**64.

Closes: #155